+++ components/browser/toolbar/src/test/java/mozilla/components/browser/toolbar/BrowserToolbarTest.kt
@@ -8,6 +8,8 @@ import android.view.View
 import mozilla.components.browser.menu.BrowserMenuBuilder
 import mozilla.components.browser.toolbar.display.DisplayToolbar
 import mozilla.components.browser.toolbar.edit.EditToolbar
+import mozilla.components.concept.toolbar.Toolbar
+import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
@@ -202,6 +204,7 @@ class BrowserToolbarTest {
         assertEquals(56, toolbar.editToolbar.measuredHeight)
     }
 
+    @Test
     fun `toolbar will switch back to display mode after an URL has been entered`() {
         val toolbar = BrowserToolbar(RuntimeEnvironment.application)
         toolbar.editMode()
@@ -285,6 +288,7 @@ class BrowserToolbarTest {
         verify(displayToolbar).addPageAction(action)
     }
 
+    @Test
     fun `URL update does not override search terms in edit mode`() {
         val toolbar = BrowserToolbar(RuntimeEnvironment.application)
         val displayToolbar = mock(DisplayToolbar::class.java)
@@ -294,9 +298,9 @@ class BrowserToolbarTest {
         toolbar.editToolbar = editToolbar
 
         toolbar.setSearchTerms("mozilla android")
-        toolbar.url = "https://www.mozilla.org"
+        toolbar.url = "https://www.mozilla.com"
         toolbar.editMode()
-        verify(displayToolbar).updateUrl("https://www.mozilla.org")
+        verify(displayToolbar).updateUrl("https://www.mozilla.com")
         verify(editToolbar).updateUrl("mozilla android")
 
         toolbar.setSearchTerms("")
@@ -417,5 +421,53 @@ class BrowserToolbarTest {
 
             verify(it).layout(50, 20, 940, 185)
         }
+    }
+
+    @Test
+    fun `editListener is set on EditToolbar`() {
+        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        assertNull(toolbar.editToolbar.editListener)
+
+        val listener: Toolbar.OnEditListener = mock()
+        toolbar.setOnEditListener(listener)
+
+        assertEquals(listener, toolbar.editToolbar.editListener)
+    }
+
+    @Test
+    fun `editListener is invoked when switching between modes`() {
+        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+
+        val listener: Toolbar.OnEditListener = mock()
+        toolbar.setOnEditListener(listener)
+
+        toolbar.editMode()
+
+        verify(listener).onStartEditing()
+        verifyNoMoreInteractions(listener)
+
+        toolbar.displayMode()
+
+        verify(listener).onStopEditing()
+        verifyNoMoreInteractions(listener)
+    }
+
+    @Test
+    fun `editListener is invoked when text changes`() {
+        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+
+        val listener: Toolbar.OnEditListener = mock()
+        toolbar.setOnEditListener(listener)
+
+        toolbar.editToolbar.urlView.onAttachedToWindow()
+
+        toolbar.editMode()
+
+        toolbar.editToolbar.urlView.setText("Hello")
+        toolbar.editToolbar.urlView.setText("Hello World")
+
+        verify(listener).onStartEditing()
+        verify(listener).onTextChanged("Hello")
+        verify(listener).onTextChanged("Hello World")
     }
 }